Refactor and override field methods for obs_instrument_coiss in ObsCassiniCommonPDS3 and ObsCassiniCommonPDS4 classes, enhancing data retrieval and ensuring consistent handling of observation fields across PDS3 and PDS4 formats.
